feat(接口): 主頁搜索接口完善+註冊接口參數調整

+++ housekeeping-admin/housekeeping-admin-service/src/main/java/com/housekeeping/admin/service/IAddressCodingService.java
@@ -13,7 +13,7 @@ import java.io.IOException;
 public interface IAddressCodingService {
 
     R addressCoding(String address);
-    Integer getInstanceByPointByWalking(String latitude1,
+    Double getInstanceByPointByWalking(String latitude1,
                                   String longitude1,
                                   String latitude2,
                                   String longitude2);
